Rename datapull to flowbyactivity
