+++ flowsa/USGS_WU_Coef.py
@@ -3,7 +3,7 @@
 # coding=utf-8
 
 """
-This script is run on it's own, not through datapull.py, as data pulled from csv in flowsa.
+This script is run on it's own, not through flowbyactivity.py, as data pulled from csv in flowsa.
 
 NOAA fisheries data obtained from: USGS Publication (Lovelace, 2005)
 
@@ -12,7 +12,7 @@ Data output saved as csv, retaining assigned file name "USGS_WU_Coef_Raw.csv"
 
 from flowsa.common import *
 import pandas as pd
-from flowsa.datapull import store_flowbyactivity
+from flowsa.flowbyactivity import store_flowbyactivity
 from flowsa.flowbyfunctions import add_missing_flow_by_fields
